Automate building, running, and linting
Most of these tasks can be handled automatically under Xcode but, until
I'm using Xcode as my editor I'm not going to bother adding a
`.xcodeproj` to this project (probably). Given that, it'd be nice to be
able to build, run, lint (and test) everything trivially and with
minimal repetition. This will also nicely document the build process,
rather than it being only in my head.