Print names of selected items in Photos
[ScriptingBridge](https://developer.apple.com/reference/scriptingbridge)
allows us to interact with apps' scripting interfaces as AppleScript
does. This gives us a (somewhat) supported way to interact with apps
and, ultimately, their data.

This configures the most basic possible protocols for interacting with
Photos itself. They allow us to get its current selection and the names
of individual “media items” (photos or videos). Given that, we can loop
over that selection and print the names.

We could work without these protocols, but it would be [more
cumbersome](https://majestysoftware.wordpress.com/2015/04/04/swift-scripting-part-2/),
involving lots of downcasting of the myriad `AnyObject` instances that
ScriptingBridge provides.