Make it possible to test the codebase
It's not possible to make an executable a test target (probably because
linking against one isn't possible?), so we should put the majority of
our functionality into a framework to allow it to be tested. This also
requires moving our main.swift into it's own target and configuring it
to depend on the framework.

We're also making the PhotosApplication and PhotosMediaItem protocols
public, so they can be accessed in the main executable, and adding a
test target to the Makefile for ease of running the (future) tests.