Make URLSession mockable
To safely (and quickly) test things that will ultimately talk to the
web, we need to be able to use our own mock of URLSession that offers
the information we want to test without risking action being taken on
live services.

To that end, we're [extending URLSession to make it conform to our new
URLSessionProtocol](http://masilotti.com/testing-nsurlsession-input/),
and creating a simple MockURLSession for use in future tests.