Add the beginnings of a FlickrAPI client
We're going to want a convenient way to make lots of calls to flickr's
API, so that should be wrapped up in a class of its own. To start with,
it will need an API key to authenticate against the API.

`call()` is a bit nonsensical at this point. Ultimately, it'll be more
like `call(method:,withParameters:?)`.