Extract queryItems collation to clarify the tests
Passing the URL's querystring through URLComponents to avoid manually
parsing it is repetitive, and makes it look like the test is doing
something it actually isn't. Extracting that to a helper clarifies that
we're just getting the queryItems, not changing anything.

I also considered wrapping this in an extension to URL that added a
`.queryItems` getter. Until I need that in the application itself, that
seems like it obfuscates what the app code is doing and what the tests
are doing (even if it was an extension in the test code, not the
app code).