Allow methods to be called with parameters
To actually be useful, we need to be able to pass parameters to the
flickr API.

This change replaces the URL generation with a more flexible approach
using `URLComponents`, which allows easier addition of arbitrary
parameters to the URL (via `URLComponents.queryItems`). It also hides
the complexity of creating an array of `URLQueryItems` from
the consumer.

I tried to make it possible to directly initialise
`URLComponents.queryItems` from a dictionary literal (further hiding
`URLQueryItem` initialising). This involved extending `Array` to
implement `ExpressibleByDictionaryLiteral`, but was problematic because
it's not yet possible to have generic types conditionally extended
depending on some criteria (this is [coming in Swift 4
](https://github.com/apple/swift-evolution/blob/master/proposals/0143-conditional-conformances.md)).
My solution involved runtime type-checking, and couldn't `throw` when
used with an invalidly-typed array (because that didn't fit the
definition of `ExpressibleByDictionaryLiteral`).

So that leaves us to `append(contentsOf:)` a map to our
required `queryItems`, as I do here.