Ensure the flickr API only returns JSON
We only ever want JSON to be returned, so we should specify that. By
default, flickr [wraps its JSON responses in a callback function
](https://www.flickr.com/services/api/response.json.html#yui_3_11_0_1_1496517335733_313).
We can't parse that, so we need to pass `nojsoncallback=1` in our
query parameters.

This commit also fixes the capitalisation of `rootURL`, and drops
`FlickrAPI.apiKey` in favour of baking it into an array of core query
parameters (including the JSON configuration mentioned above) that we
want to send for every API request.