Ensure make always builds when sourcefiles change
Because make's file globbing only works for a single directory, we need
to use find to get all the files in subdirectories. I tried to set up
targets for each of the build artefacts, but it's not clear to me which
files constitute the useful artefacts of the library, so this is
simpler (and works well enough).

This also adds an `all` make target, to build, lint, and test, removes
some of the less useful dependencies from `lint` and `test`, and
removes the as-yet-unused cleanup of a hypothetical `Packages`
directory.