Prettify test output using xcpretty if possible
If we have [xcpretty](https://github.com/supermarin/xcpretty/)
installed, we should use it to more nicely format the output.

Unfortunately, if `swift test` auto-builds, those steps are hidden by
xcpretty because they don't match any of its expected build steps. So
it's a good job we force them to happen before running tests.

Of note here is the fact that have to redirect STDERR to STDOUT because
apparently `swift test` outputs to STDERR for some reason?