Require an API key and instantiate FlickrAPI
To do anything with the flickr API, we need an API key, so fail without
that set as an environment variable. We could probably accept the API
key as a command line parameter, but that then forces it to end up in
the user's command history. This way, they could export it (or
something) if they were particularly paranoid. Also this is easy(ish).

To make printing to `FileHandle.standardError` easier, we're also
extending `FileHandle` to conform to `TextOutputStream` so it can be
used with `print(_:,to:)`. I can't work out how to test this easily so,
for now, I'm not going to bother. I tried writing to tempfiles and
reading them back and so on, but that didn't seem to work.