Simplistically find candidates from flickr by date
It seems likely that the photos on flickr will have the same date/time
as the local copy, so let's see if we can find them. This does, of
course, require us to be able to get the date for each
`PhotosMediaItem`.

This currently assumes you're running the script as me, which should
probably change in future.