Attempt to confirm flickr matches by name
The photos we get back from the flickr search aren't guaranteed to be
the ones we want, so we can simplistically check by filtering the
candidates based on their name.

This might result in no matches, or in many, so we should flag those
early and abort. Where we only have one match, we can (probably?)
assume that's the one we want: the date and title are the same, so we
can start pulling data from it to our local copy… maybe.