Add the beginnings of location comparison logic
If there's a location on flickr, we should either use it (if the photo
has none) or offer the choice to the user (if the photo has a
location). If there's none, show any tags flickr has in case they can
offer a clue to what the location should be.

In future, most of this logic should probably be embedded into
PhotosMediaItem (and tested!), but that can come later.

This also pulls the `standardError` file handle out to be available
across the entire main file.