Remove dependency on SwiftyJSON
SwiftyJSON is nice, but it ends up having to be littered throughout the
codebase. Removing it makes things a bit harder now, but allows us to
do things in [a more Swift-native way
](https://developer.apple.com/swift/blog/?id=37) later. Further,
[Swift 4's`Codable` types
](https://github.com/apple/swift-evolution/blob/master/proposals/0167-swift-encoders.md)
will actually make this even easier. Given we're on Swift 3, though,
this is what we have.

Because the native types are optional, I've littered the code with a
few guards to keep things easier.

It's not clear to me why `setLocation` in
`Sources/photos-metadata-fixer/main.swift` requires the latitude and
longitudes to be cast first to `String`s then use those to create
`Double`s; other minimal test cases I've made don't exhibit that
behaviour. Something is awry there, but I'm not sure what. They appear
to be ending up as NSTaggedPointerStrings… whatever they are.

This also squashes a compiler warnings about `var newLocation` not
being written to; that will change revert in future. For now, it can be
this way.