Use native photo search function
Now we have it, we should use it. Because `searchForPhotos()` returns
`FlickrPhoto`s, that also means we don't have to walk the JSON tree and
directly access the instance variables.

It also turns out I forgot to make `searchForPhotos()` and
`FlickrPhoto`'s instance variables public, so we should do that too.