Simplify test assertions on containers
This adds `assert(_:, contains:)`, which more simply shows what we're
actually testing for (x contains y), and ensures we return a consistent
error message from it.

Of note, here, we're using the `#file` and `#line` [literal expressions
](https://developer.apple.com/library/content/documentation/Swift/Conceptual/Swift_Programming_Language/Expressions.html#//apple_ref/doc/uid/TP40014097-CH32-ID390)
to ensure any test failures point to the *actual* originating line and
file, rather than the line calling `XCTAssertTrue`.

We should possibly be using
[`recordFailure(withDescription:inFile:atLine:expected:)`
](https://developer.apple.com/documentation/xctest/xctestcase/1496269-recordfailure)
for this, but this way seems to most clearly signpost what the actual
intent is (and keeps this as a very simple, clear wrapper).