Parse real flickr photo JSON
It turns out that flickr's latitudes and longitudes are strings, not
doubles, so we need to cast them through strings. This seems like a
slightly cumbersome way to do that, but I can't currently think of a
cleaner way given `Double.init()` only accepts `String`, not `String?`.

For photo info responses, flickr also moves the title from being the
value of the `title` subscript to a sub-element called `_content`
(because their JSON is a shitty conversion from the XML).

This commit also slightly restructures the location test.