Find more photos by brute-forcing timezones
If we already knew the photos' location, we could use that to work out
what the correct timezone would be for those whose timezone is incorrect.
As we don't, we can't.

This gets a list of all the timezones, works out the unique GMT offsets
among those (because, for example, America/New_York and America/Lima
are the same offset), and tries to match the photo at each of those
unique offsets.

Because this is making *lots* (~40) more API requests per photo, it
makes things significantly slower, but it successfully matches ~365
more images in my sample of ~2460, with no increase in excessive
matches.