Set dates on FlickrPhotos
We need access to the dates on photos. Flickr can return them either
from the getInfo API per photo, or in the search results by passing
`extra=date_taken`. Unfortunately, flickr puts the date taken in
different places in the response depending on which response it is, so
we have to vary based on that.

For some reason, [flickr's dates are passed around in MySQL's datetime
format](https://www.flickr.com/services/api/misc.dates.html), so we
need to use a custom formatter to parse them. [Apple suggests setting
the locale to a POSIX locale](https://developer.apple.com/documentation/foundation/dateformatter#2528261)
when working with fixed format date representations like this.