Set FlickrPhotos' location from search responses
As with dates, flickr moves the location around in the JSON response
depending on which API method we're using, so we need to account for
that.

This also renames the other location test to be a little clearer,
removes a couple of unnecessary assertions, and updates the tags test
to match.